simplesim: Add session id to GET_INFO messages
