+++ softwareComponents/simplesimLib/include/simplesim/distributor.hpp
@@ -66,7 +66,7 @@ private:
         onRequest( *reqCopy );
     }
 
-    rofi::messages::DistributorResp onGetInfoReq();
+    rofi::messages::DistributorResp onGetInfoReq( SessionId sessionId );
     rofi::messages::DistributorResp onLockOneReq( SessionId sessionId );
     rofi::messages::DistributorResp onTryLockReq( ModuleId moduleId, SessionId sessionId );
     rofi::messages::DistributorResp onUnlockReq( ModuleId moduleId, SessionId sessionId );
